Moved packages under pkg directory
Moved model, compile, validate, and taxonomyio.
